cleaned up docs
